Over clever.
Feel free to delete this. It is the notion that if `write` returns
`false` we don't really have to wait immediately for a drain. We can
wait until we come back around again, if the write hasn't drained in the
meantime.

However, we don't wait until we are over the high-water mark. If we are
writing huge buffers, then yes, we might have to wait, but I can't
really think of a case where we're not delaying the inevitable. It would
be worth it, at some point, when I have an application that uses
Staccato, to test to see if something like this would be faster, or at
least to see if it would abvoid draining.

At long last, I'm finding that premature optimization creates more
problems than it is worth.