Enlightened, simplified Staccato.
Error handling from streams is difficult because it is not consistent.
Sockets are especially difficult. It was my hope that Staccato would
turn any socket error into an exception you could catch, but after all
this time, I'm still not sure when a socket is going to raise an error.

Plus, it is makes it difficult if Staccato is supposed to be your safety
to shut things down in the right order. You'd need to leave Staccato
lying around to pick up on exception raised even you're done reading or
writing for now. I'd rather register a logger on the socket and forget
about it, or maybe make some sort of scream box for sockets to cry into
when they will.

I do want an error to stop things. Reading is simpler that writing. You
generally get an error and that's all from the read side of things. If
`Staccato.Readable` is working out for an application, I've added
`raise` function that you can call to raise the `error` property if
there is one, if that's how you'd like handle it.

It is not often the case that I really want to k