Fix socket write error hang, add state property.
When you write to a closed socket the asynchronous call to
`Stream.write` returns `false` and while you where calling
`Stream.write` and error was emitted. Thus, we need to check `destroyed`
when we return from the call to `write` to see if that `false` value
means that the data was buffered or whether the write was in error.

Additionally, I added a `state` property so that I can debug these
issues, see where things are hanging.